requirements.txt
flask
requests
yfinance
google-generativeai
python-dotenv
gunicorn
